Rename generation -> extraction everywhere
