+++ .vscode/launch.json
@@ -23,7 +23,7 @@
             "name": "Python: Generation Cuda",
             "type": "python",
             "request": "launch",
-            "module": "elk.generation_main",
+            "module": "elk.extraction_main",
             "console": "integratedTerminal",
             "justMyCode": true,
             "args" : [
